hw4/requirements.txt
pystan
numpy
scikit-learn
matplotlib
word2vec
nltk
adjustText
Pillow
